pyproject.toml
[tool.poetry]
name = "cencosud"
version = "0.1.0"
description = ""
authors = ["andreslistz <andres.sandoval@alumnos.usm.cl>"]

[tool.poetry.dependencies]
python = "^3.7"
SQLAlchemy = "^1.3.20"
python-dotenv = "^0.15.0"
pydantic = "^1.7.2"
pandas = "^1.1.4"
statsmodels = "^0.12.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.6.0"
jupyter = "^1.0.0"
black = "^20.8b1"
matplotlib = "^3.3.3"
numpy = "^1.19.3"
autopep8 = "^1.5.4"
docformatter = "^1.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
